+++ bitvunit-core/src/test/java/de/codescape/bitvunit/rule/page/SingleTitlePerPageRuleTest.java
@@ -9,40 +9,37 @@ import static de.codescape.bitvunit.util.html.HtmlPageUtil.toHtmlPage;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-public class TitleForPageRuleTest {
+public class SingleTitlePerPageRuleTest {
 
-    private TitleForPageRule rule = new TitleForPageRule();
+    private SingleTitlePerPageRule rule = new SingleTitlePerPageRule();
 
     @Test
     public void ruleHasAName() throws Exception {
         assertNotNull(rule.getName());
-        assertEquals("TitleForPage", rule.getName());
+        assertEquals("SingleTitlePerPage", rule.getName());
     }
 
     @Test
-    public void noTitleTagPresent() throws Exception {
-        String content = "<html><head></head></html>";
+    public void multipleTitlesOnPage() throws Exception {
+        String content = "<html><head>" +
+                "   <title>First</title>" +
+                "   <title>Second</title>" +
+                "   <title>Third</title>" +
+                "</head><body><p>text</p></body></html>";
+        Violations violations = rule.applyTo(toHtmlPage(content));
+        assertViolations(violations, rule, 3);
+    }
+
+    @Test
+    public void missingTitleOnPage() throws Exception {
+        String content = "<html><head></head><body><p>text</p></body></html>";
         Violations violations = rule.applyTo(toHtmlPage(content));
         assertViolations(violations, rule, 1);
     }
 
     @Test
-    public void emptyTitleTagPresent() throws Exception {
-        String content = "<html><head><title></title></head></html>";
-        Violations violations = rule.applyTo(toHtmlPage(content));
-        assertViolations(violations, rule, 1);
-    }
-
-    @Test
-    public void whitespaceOnlyTitleTagPresent() throws Exception {
-        String content = "<html><head><title>   </title></head></html>";
-        Violations violations = rule.applyTo(toHtmlPage(content));
-        assertViolations(violations, rule, 1);
-    }
-
-    @Test
-    public void noViolationRaisedByRule() throws Exception {
-        String content = "<html><head><title>Hello World</title></head></html>";
+    public void exactlyOneTitleOnPage() throws Exception {
+        String content = "<html><head><title>Good title</title></head><body><p>text</p></body></html>";
         Violations violations = rule.applyTo(toHtmlPage(content));
         assertNoViolations(violations, rule);
     }